requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langgraph
langchain
langchain-openai
python-dotenv==1.0.0
numpy==1.24.3


supabase
gotrue
